refactor(llm): move FriendlyError from discord to llm
